iot-factory-cdk/requirements.txt
aws-cdk-lib==2.95.1
constructs==10.2.69
